requirements.txt
absl-py==2.1.0
astunparse==1.6.3
biopython==1.66
cachetools==5.5.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
cvxopt==1.2.7
cycler==0.11.0
Cython==3.0.11
decorator==4.4.2
dgl-cu101==0.6.1
dill==0.2.9
dm-tree==0.1.8
docker-pycreds==0.4.0
eden==0.2
eden-kernel==0.0.0
flatbuffers==24.3.25
fonttools==4.38.0
future==1.0.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grakel==0.1.8
grpcio==1.62.3
h5py==3.8.0
idna==3.10
importlib-metadata==6.7.0
Jinja2==3.1.4
joblib==1.3.2
keras==2.11.0
Kivy-Garden==0.1.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
mpmath==1.3.0
networkx==2.5.1
nose==1.3.7
numpy==1.21.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
platformdirs==4.0.0
powerlaw==1.5
prettytable==3.7.0
protobuf==3.19.6
psutil==5.8.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyemd==0.5.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==0.20.4
scipy==1.4.1
sentry-sdk==2.15.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.10.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-gan==2.1.0
tensorflow-hub==0.16.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-probability==0.19.0
termcolor==2.3.0
threadpoolctl==3.1.0
toolz==0.12.1
torch==1.8.1
torch-cluster 
torch-geometric==2.0.4
torch-scatter 
torch-sparse 
torch-spline-conv 
torchvision==0.9.1
tqdm==4.66.5
typing_extensions==4.7.1
urllib3==2.0.7
wandb==0.18.3
wcwidth==0.2.13
Werkzeug==2.2.3
wrapt==1.16.0
zipp==3.15.0
